Changing the data type

+++ homework.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass, asdict
-from typing import Dict, Type
 
 
 @dataclass
@@ -153,7 +152,7 @@ def main(training: Training) -> None:
 
 
 if __name__ == '__main__':
-    packages: list[str, list[int]] = [
+    packages: list[str, list[float]] = [
         ('SWM', [720, 1, 80, 25, 40]),
         ('RUN', [15000, 1, 75]),
         ('WLK', [9000, 1, 75, 180]),
